+++ Cecilifier.Ide.Plugin/protocol/src/main/kotlin/model/rider/Cecilifier.Ide.PluginModel.kt
@@ -1,0 +1,32 @@
+package model.rider
+
+import com.jetbrains.rider.model.nova.ide.SolutionModel
+import com.jetbrains.rd.generator.nova.*
+import com.jetbrains.rd.generator.nova.PredefinedType.*
+
+@Suppress("unused")
+object Cecilifier.Ide.PluginModel : Ext(SolutionModel.Solution) {
+
+    val MyEnum = enum {
+        +"FirstValue"
+        +"SecondValue"
+    }
+
+    val MyStructure = structdef {
+        field("projectFile", string)
+        field("target", string)
+    }
+
+    init {
+        //setting(CSharp50Generator.Namespace, "ReSharperPlugin.Cecilifier.Ide.Plugin.Rider.Model")
+        //setting(Kotlin11Generator.Namespace, "com.jetbrains.rider.cecilifier.ide.plugin.model")
+
+        property("myString", string)
+        property("myBool", bool)
+        property("myEnum", MyEnum.nullable)
+
+        map("data", string, string)
+
+        signal("myStructure", MyStructure)
+    }
+}