+++ Cecilifier.Ide.Plugin/src/dotnet/ReSharperPlugin.Cecilifier.Ide.Plugin.Tests/TestEnvironment.cs
@@ -1,0 +1,21 @@
+﻿using System.Threading;
+using JetBrains.Application.BuildScript.Application.Zones;
+using JetBrains.ReSharper.TestFramework;
+using JetBrains.TestFramework;
+using JetBrains.TestFramework.Application.Zones;
+using NUnit.Framework;
+
+[assembly: Apartment(ApartmentState.STA)]
+
+namespace ReSharperPlugin.Cecilifier.Ide.Plugin.Tests
+{
+  [ZoneDefinition]
+  public interface ICecilifierIdePluginTestZone : ITestsEnvZone, IRequire<PsiFeatureTestZone>
+  {
+  }
+
+  [SetUpFixture]
+  public class TestEnvironment : ExtensionTestEnvironmentAssembly<ICecilifierIdePluginTestZone>
+  {
+  }
+}