+++ Cecilifier.Ide.Plugin/src/dotnet/ReSharperPlugin.Cecilifier.Ide.Plugin/SampleComponent.cs
@@ -1,0 +1,19 @@
+﻿using JetBrains.Application;
+using JetBrains.Lifetimes;
+using JetBrains.ReSharper.Feature.Services.QuickFixes;
+using JetBrains.ReSharper.Intentions.CSharp.QuickFixes;
+
+namespace ReSharperPlugin.Cecilifier.Ide.Plugin
+{
+    [ShellComponent]
+    internal class SampleQuickFixRegistrarComponent
+    {
+        public SampleQuickFixRegistrarComponent(IQuickFixes table)
+        {
+            table.RegisterQuickFix<SampleHighlighting>(
+                Lifetime.Eternal,
+                h => new SampleFix(h.Declaration),
+                typeof(SampleFix));
+        }
+    }
+}