+++ Cecilifier.Ide.Plugin/src/dotnet/ReSharperPlugin.Cecilifier.Ide.Plugin/SampleSettings.cs
@@ -1,0 +1,15 @@
+using JetBrains.Application.Settings;
+using JetBrains.Application.Settings.WellKnownRootKeys;
+
+namespace ReSharperPlugin.Cecilifier.Ide.Plugin
+{
+    [SettingsKey(
+        typeof(EnvironmentSettings),
+//        typeof(CodeEditingSettings),
+        "Settings for Cecilifier.Ide.Plugin")]
+    public class SampleSettings
+    {
+        [SettingsEntry(DefaultValue: "<default>", Description: "Sample Description")]
+        public string SampleText;
+    }
+}