+++ Cecilifier.Ide.Plugin/src/rider/main/kotlin/com/jetbrains/rider/plugins/sampleplugin/CecilifierAction.kt
@@ -1,0 +1,8 @@
+package com.jetbrains.rider.plugins.cecilifier.ide.plugin
+
+import com.jetbrains.rider.actions.base.RiderAnAction
+
+class CecilifierAction : RiderAnAction(
+        "CecilifierAction",
+        "RunCecilifier",
+        "Run cecilifier on current document.")