+++ Cecilifier.Ide.Plugin/src/rider/main/kotlin/com/jetbrains/rider/plugins/sampleplugin/SampleOptionsPage.kt
@@ -1,0 +1,9 @@
+package com.jetbrains.rider.plugins.cecilifier.ide.plugin
+
+import com.jetbrains.rider.settings.simple.SimpleOptionsPage
+
+class SampleOptionsPage : SimpleOptionsPage("Sample Options", "SampleOptionsPage") {
+    override fun getId(): String {
+        return "SampleOptionsPage"
+    }
+}